Coolo: new libtool is out (0.9f) It has some new features: * Added new `--disable-static' flag to disable building static   libraries on platforms that have shared libs.  (This one will please Kalle) * New `--features' flag to display configured libtool attributes. * Support *-*-linux-gnu* as an alias for *-*-linux*. * Support for *-*-openbsd* and *-*-freebsd3*.
And some more, e.g. dl support.

I've subscribed to the libtool and the automake (Btw: checkout 1.1p on fiwi02)
mailing list, so we have always the newest. I found already two bugs and per-
haps our project is the best beta test for libtool.

Perhaps this will cause some trouble on the one or the other system, but I
don't hope so.

svn path=/trunk/kdelibs/; revision=218