extended kconfigbackend so that it understands other resource types than "config".  Made it findAllResources recursively.  Shouldn't be a performance hit for type "config" which is a flat directory, but it will enable use of resources like "apps" successfully.
Extended the kdesktopfile class so that it does config "shadowing" and
"merging" like a regular KConfig.  It is now a thin wrapper on top of
KConfig instead of KSimpleConfig, which simply disables kdeglobals
parsing, sets the resource type to "apps" instead of "config", sets the
desktop group as default, and then provides some convenience methods.

svn path=/trunk/kdelibs/; revision=26511