Here comes the new DialogBase which is renamed to KDialogBase. The old DialogBase remains in the lib for now, until I have changed the code where it is used.
The KJanusWidget is used inside KDialogBase but can also be used as a
"standalone" widget. The KAboutDialog has already been changed to use
KDialogBase.

Makefile.am is changed in order to add the two new classes. I think the
old DialogBase can be removed quite soon. It is only used in kdelibs/kdetest
and in kdelibs/kab. (in cvs of course ;) The DialogBase and KDialogBase are
more or less compatible so a change wont be difficult. I will commit
changes for the code that need to be altered on Monday evening.

svn path=/trunk/kdelibs/; revision=29712