moved the new html implementation over to the main branch.
BEWARE: It might not compile. I'll fix this in the evening, after I
come home from Erlangen. It'll break apps using khtml. As I will do
some rewrite of the external API in the next one or two days, it's
probably not worth fixing that...

But we have JAVA now!

svn path=/trunk/kdelibs/; revision=30525