By default, toolbars will no longer honor the user-set icon mode (icons only, text under icon, etc).  This is for the benefit of apps that have many many toolbars -- multiple toolbars look very bad in any other mode but "icons only".
There *ARE* two ways to force a mode change, though.  The first is a
boolean flag in the constructor (honor_mode -- default 'false') that,
if set to true, will allow the toolbar to read in the global settings.
It is up to the application OR more importantly, the application
framework to decide which toolbar should honor the setting.  This is
the "main" toolbar, in 99.9% of all apps.  In current apps, this is
done with 'toolBar(0)' in KTMainWindow or the "mainToolBar" toolbar in
the KParts realm.

I also added a sub-menu to the toolbar context menu.  This will allow
users to change the style for individual toolbars "on the fly"!

svn path=/trunk/kdelibs/; revision=37067