Changed the way kiconloader handles toolbar icons... but actually managed to preserve the "old" way of doing this just in case!
The "new" toolbar dirs will be integrated (sometime soon) into the
current icons tree.  Getting a toolbar icon can now be done with
something like: iconLoader()->loadApplicationIcon("toolbar/filenew");

However, I left all of the old toolbar loading code in there.. and
added an emergency backup procedure in loadApplicationIcon that will
use the old code to try and find the icon if all else fails.

The end result is that the code will find the the toolbar icon pretty
much no matter where it is.

svn path=/trunk/kdelibs/; revision=37227