Hmm.  I love how things get broken and or mangey.  Basically I fixed IPv6 support (now that I've got two boxes with IPv6 stacks..).  However, in the process I renamed KSocket*::addr() to ::ipv4_addr.  I know I didn't add that function, and don't know whose moronic idea it was.  Grr.  Anyways I also unconditionalized a few includes, because they were being included elsewhere in an unconditionalized statement, either in ksock.cpp or elsewhere in kdelibs.  I don't know if this will break Linux builds, but I can only assume it won't because of the nature of the changes, and I don't know if this will work on other IPv6 staks (I'm using FreeBSD-current's stack, which is essentially the KAME stack), it should because I essentially followed rfc2133 (ipv6.org).  Next comes some enums and a method to frob the IPv6 goodies and then comes konsole.
The way it works is, if you compile it without defining INET6, it should
work like before.  With INET6 it should work like before (but not be
entirely binary compatible, I think), but when you pass it a hostname it
will look for AAAA records (IPv6 hostnames) first, then A records (IPv4).
If you pass it an "IPv6 ip address" it will work, an IPv4 mapped onto an
IPv6 should work too, an IPv4 won't (didn't try without INET6 defined).
This should be easy to fix.  It's quite seamless, but I think that adding
a "family" argument woudln't hurt (for instance kame.net had an ipv4 and
ipv6 "box" for debugging).  If KURL would allow numbers in the protocol
qualifier, this would be useful so one could do: http6://www.kame.net and
get the IPv6 version.

svn path=/trunk/kdelibs/; revision=38619