Ok, here's some "fixes" that allow reasonably transparent IPv6 support.  If you are a consumer of initSockaddr, all you have to do is use ksockaddr_in instead of struct sockaddr_in and then use the appropiate macros instead of accessing the members directly.  I might end up doing a class to properly encapsulate this whole mess and this would allow for easier copying of hostents and the like.
As it stands it includes only the IPv4 headers and then checks to see if
PF_INET6 is defined.  If it's defined IPv6 stuff is used.  This means that
if your IPv4 headers don't actually include the new stuff you're somewhat
dead in the water (if this happens, a safer bet would be probably to test
for a generic IPv6 macro like IN6_IS_ADDR_V4COMPAT).

The following commits will allow kio_http and kio_ftp to correctly access
IPv6 hosts (try it out, www.freenet6.net).

However, due to the type saftey (read: name mangling) this is not really
binary compatible, and as always if you switch around your headers and
don't recompile kdelibs, weird things should result.

Hah, I'm off to see the wizard, the wonderful wizard of oz.

svn path=/trunk/kdelibs/; revision=38685