Added library closures for -frepo compilation (experimental). If you to compile with -frepo, do in each directory that contains a library:
make closure all

This ensures that all templates are properly instantiated before the
library is linked. (The trick is to compile a small application that
links all object files, which causes collect to instantiate all templates
that are required). Since -frepo is buggy in quite some amount of gcc
versions, this may not work for you.

svn path=/trunk/kdelibs/; revision=38925