Implemented handling of modified read-write parts :
ReadWritePart::closeURL() asks for saving, with yes/no/cancel.

Destroying a part doesn't do that (too late for cancelling) so the
host application should really call closeURL first.

setModified turned into a slot (no arg anymore) to make it possible
to connect it to e.g. textChanged().

The notepadpart tests all this, and works fine.
Warning, it opens the .rc file, so be careful (saving works as well!)

svn path=/trunk/kdelibs/; revision=39087