Changed the handling of the window caption : now a signal emitted by the part, connected/disconnected by the mainwindow. A default implementation of guiActivateEvent() has been added to KReadOnlyPart to set the caption to m_url - this is what we want in 99% of the cases.
Added signal for setting a statusbar text as well.
This means the shell handles the statusbar now, it's not a per-part thing.
NOTE : shells that embed part that emit setStatusBarText should have a
StatusBar tag in their XML.
mainwindow provides a default implementation (calling message()) but apps
can override it if they have a complex statusbar (konqy does).

svn path=/trunk/kdelibs/; revision=40037