Two changes:
1) Added private data class pointers for all action classes.  ALL KDE
   classes need these before 2.0 so we might as well start adding them
   now.

2) Added additional constructors to all of the actions that permit the
   developer to supply only the name of an icon.. not the icon itself.

That last one needs some explaining, I think.  The problem I (and
others) was running into was that containers may want different sizes (and
maybe later, colors) for their icons.  However, when you pass along the
icon itself, the containers were limited to simply resizing the icon --
not exactly a good way to handle it.

So we now have the option to pass along just the name.  This will assume
that the container knows how load an display its own icon.  Those
containers that don't know how will be fed a small, default-color
toolbar icon.

This makes some sort of sense sometimes.  That is, ktoolbar will handle
its own icons sooooon (it works on my hard drive) and qpopupmenu always
wants small.  Since those two are 99% of all uses of KAction (with
icons), I think we are mostly covered.

svn path=/trunk/kdelibs/; revision=42285