Reverted parts of my commit (too bad there is no canonicalPath in QFileInfo)
GOSH. There was an awful confusion in KIO::findDeviceMountPoint.
Stephan: your code looked like it tried to match a DIRECTORY name with
a mountpoint - to know which partition we are on. This could be useful
somewhere else, but KIO::findDeviceMountPoint is for something different:
it's "give me the mount point for device /dev/abcd". That's a lot more
straight forward ! (Note the comparison happened on FSNAME(me), so it's useless
to do partial string matching here)

I guess we should write also KIO::findMountPointFromDirectory or
findDeviceFromDirectory, with the old code. Any use for that ?
(With this naming scheme, the current one is findMountPointFromDevice...)

svn path=/trunk/kdelibs/; revision=42508