Two changes: o made actionCollection() virtual.. just in case a XML GUI client might   want to supply an action collection NOT created by declaring the   collection as an action's parent. o added support for a file-wide "nomerge" tag.  if it is found, then   the mergeXML code will NOT merge the local file with the global one.   rather, it will take the local file as the finished product and go   with that
That last change enables us to switch around standard entries on an
app by app basis.  For instance, it is now possible to have the File
menu on the extreme right (instead of left) for only one app... that
was not possible before.

The downside to all that is that this particular app will no longer
catch any global changes to the UI.. oh well.

A container-specific 'nomerge' tag will be coming as soon as I need
it.

svn path=/trunk/kdelibs/; revision=42815