This change won't make much sense now but will in a day or two when I check in some new code.
Previously, if setXMLFile couldn't find the given XML file (e.g., it was
lost or didn't exist yet), the function returned an error.  I changed
it so that it instead calls 'setXML(QString::null)'.  That call, of
course, means that the merging code should simply work with the global
standards file.

This change will be necessary Real Soon Now(tm).  Trust Me(tm)

svn path=/trunk/kdelibs/; revision=42907