Two commits in one...
The main commit is the addition of the KEditToolbar dialog.  This will
allow users to dynamically edit their toolbars.. at least in any app
that uses the new XML UI framework.

It currently is about 50% implemented, but does work passably on such
apps as kedit.  It doesn't work very well on KParts apps.. yet

I also changed the createGUI behavior just a tad.  Now, passing along
a null string does not mean to just use the global resource file.  It
also creates a default local file in case you want to add local toolbars
there later.  Actually, it just creates the name.. the file itself isn't
created until something is put there.

All this will be discussed in a little more detail on kde-core-devel

The second commit is some prototypical code for an indicator widget as
discussed on kde-core-devel.  It somewhat works but not perfectly, yet.
If the setIndicatorWidget function is never called, then nothing to do
with it will ever happen.  e.g., no current apps will be affected at
all.

svn path=/trunk/kdelibs/; revision=42995