Two changes:
o We no longer hard code the toolbar button font.  Rather, we use the
  global KGlobal::buttonFont() method to get whatever the user set.  This
  is pretty cool.  Now we just need to be able to set the color of the
  highlight!

o Disabled buttons no longer have greyed out text UNLESS the button is
  text only

That last one probably needs some explaining.  I got an email from a person
with not-too-great-eyesight (a very very common occurence in the geek
world) saying that it is difficult to see what the buttons are when they
are disabled.  The person had a good point.  It's quite enough to see that
the button itself is disabled by the disabled pixmap... we don't have to
make the button hard to read, though!

Hard-coded styles like B2 will not honor that change, so I'll be switching
them by hand, soon.

svn path=/trunk/kdelibs/; revision=43026