FOUND IT ! Fixed konqueror "Double QObject deletion" !!!
Dawit !!! If you do a setPopup() and make a local copy, do NOT destroy
it in the destructor ! KToolbarButton should NOT take ownership of something
it hasn't created...

Whoever creates the popup should destroy it.
konqueror does, obviously (hence the crash), but I don't know about other
apps using this feature, and possibly not destroying the popupmenu.

svn path=/trunk/kdelibs/; revision=43031