Hmm.. that was almost too easy.
Anyway, I fixed the annoying bug where you could not merge any
non-standard actions into any standard menu after bookmarks.  AFAIK,
the only commited app that fell for this was kmines but I have been
getting pretty steady bug reports in my mailbox from authors in the
process of converting.

The fix was simple: before I actually do a merge, I first check to see
if the local container name appears in the global file later on.  If it
does, then I postpone the merge for that container as it will be
handled later on.  Simple!

svn path=/trunk/kdelibs/; revision=43059