Added "void KDialogBase::setIconListAllVisible( bool state )" and "void KJanusWidget::setIconListAllVisible( bool state )"
This function will prevent the vertical scrollbar to be used in the
icon list by increasing the height of the KJanusWidget widget and/or
the KDialogBase dialog. It looks bad if a scroll bar becomes visible
just because the icon list is a few pixels too tall.

kjanuswidget.cpp+h : func( void ) -> func()

svn path=/trunk/kdelibs/; revision=43920