Added license declarations.
HOWEVER!  I don't know who the authors to all these files are.  I just
put David as the copyright on all of them even though I know that coolo
and waldo and matt koss and torben probably many others should have
their names in addition or instead of david's.

SO *please* go through the files and add your names to the copyright
line if you think you should be there.

svn path=/trunk/kdelibs/; revision=44997