Implemented icon effects.
The effects are done in KIconEffect which contains some duplicated code from
KImageEffect. I think there's not other way because KIconLoader depends on
KImageEffect and we don't want to link kdecore (KIconLoader) to kdeui
(KImageEffect).

The duplicated to is not much but I'm open to suggestions how to do this
cleaner.

Icon effects are used to hint about the state of an icon: Default, Active or
Disabled. This change is source and binary compatible.

svn path=/trunk/kdelibs/; revision=45071