Added support for the key accel chooser dialog to handle actions.  It was actually fairly trivial... partially because it doesn't entirely work :-)
You currently can't save the changes permanently and it doesn't always
read in the old saved changes correctly.  The saving part is the tricky
one... where should it be saved?  Like KAccel in config files?  Or as
action properties in the XML files.  I am favoring the latter.  To do
that, though, I may need to pass along a KXMLGUIFactory pointer instead
of the action collection...

svn path=/trunk/kdelibs/; revision=45322