most changes to the dom tree should now also trigger a repaint/relayout automatically. Harri, could you give it a try?
Frederik, IMO, the Node::applyChanges() method should not be needed at all.
I added the appropiate calls to the setXXX methods in TextImpl and
ElementImpl. Like this, the changes should get applied to the rendering
tree automatically. If you don't have any good reasons for having the
method in DOM::Node, I'd like to remove it before monday.

svn path=/trunk/kdelibs/; revision=45334