Here comes KIO's UI server, which the Observer starts using startService and talks to via DCOP. More showProgressInfo, make the server assign an ID to a job and send it back, so that the observer uses that job id when sending signals (not pointers!).
But I realise this whole thing only makes sense if the point is to use
a ListProgress or anything else (somebody said kicker applet?) showing _all_
jobs running at a time. If we end up using DefaultProgresses most of the
time, they could be in the process just as well, like we did before...
Anyway. This stuff is real fun.

Charles, wanted an example of autostarted DCOP server ? Here is a good one...

svn path=/trunk/kdelibs/; revision=45362