Slightly different API for loadIcon().
You can pass a size along with the group now. If this size is nonzero, it
will override the default size bound to the group.

I hope people will make more extensive use of groups now. If you don't want
to use groups (i.e. you want to prevent effects from being applied), you
should use KIcon::NoGroup.

I hope this will be the last change in the API of the iconloader. Apologies
for the changes so far.

svn path=/trunk/kdelibs/; revision=45441