Baby steps to reimplementing mac menu mode for KDE.  I *really* missed that from KDE 1.x
Unfortunately, this is starting to get into events and wm issues that
are beyond my realm of knowledge, so it doesn't *quite* work right.

What does work:
o If 'macStyle' is set to true, then ktmainwindow apps will have their
  menus on the top of the screen

However:
o The most recently *started* window will always have its menu on top...
  I can't figure out how to have the activated window's menu on top
o If you move the app to another desktop, the menu does not follow.  I
  think this is kdesktop or kwin specific since KDE 1.x apps exhibit
  this behavior under a KDE2 desktop as well
o The menu name ('<caption> [menu]') shows up on the kicker taskbar.  This
  also shows up with KDE 1.x apps so this is also a kwin or kicker
  issue
o This is just a bare menu.  I didn't reimplement the handle or pop-down
  system menu

This is binary compatible

svn path=/trunk/kdelibs/; revision=45830