Fixed the following:
1.) Remove all the hack done to support invalid forms of the smb protocol.
This is now properly handled by the KURIFilter class.  KURL has to remain
generic.

2.) Associated with the above work, fixed two places where the original
correct implementation was commented out and replaced by an incorrect one.
This completely broke the parser's ability to correctly recognize some
malformed URLs.  In this case it was the formats <scheme>: and <scheme>://,
without any other character(s), i.e. the URL was only "smb:" or
"smb://".  Both of these are INVALID URLs, but were being passed as valid
ones. BTW, the fix is protocol antagonistic and works just the same for all
the other protocols as well.

3.) KURL now correctly rejects URLs that only contain a scheme part.  Thus,
"http:" or "smb:" now correctly returns malformed.

4.) For whatever reason KURL was made to assume the URL was valid before it
even got a chance to parse it.  This has been changed now.  By default the
URL is assumed to be malformed until it gets parsed correctly.  Thus, the
flag is set only if the URL reaches "NodeOk:" state.

5.) Removed the assigment that re-initalized the protocol to empty qt the
beginning of the parse.  If someone needs this for whatever reason, please
change the behavior change it at the ::reset() function and do not add any
more hacks to the parser. (DA)

svn path=/trunk/kdelibs/; revision=46063