list-style-position: inside works now. Marker positioning is a bit wrong, but I'll leave that to someone else ;-)
Lists in a right-to-left context mostly work. Two problems remain.
The trivial one is to put the '.' in numbered lists on the left of the
number. The other one is getting the indentation right if the styles
directionality is RTL. At the moment these lists are still indented on
the left, instead of right. I'm not sure how to solve this.

svn path=/trunk/kdelibs/; revision=46193