Some significant speed improvements.
We set NoBackground for all widgets we paint entirely like toolbars, and use
Matthias Ettrich's cool new parent pixmap stuff instead of masks as long as
we are neither in toolbars or KHTMLView (where it won't work for now).

Lars: We were discussing masks in KHTML and how to use it while using parent
brush origin stuff in styles. Doing this should take care of it - you should
not have to do anything.

Thing seems fast to me now :)

svn path=/trunk/kdelibs/; revision=46453