Adapted highcolor optimizations from B2, and did some fixes. Then did some more fixes. Then even more... There was a lot to fix here ;-)
I usually only use B2 and B2/Highcolor, so if someone that is familiar with
KStyle (Rik or Kurt?) wants to use this for awhile to test for bugs and make
improvements that would be great. Just make sure it remains somewhat Next-like
(simple rectangular buttons with light as the highlight and solid black as
the shadow - although I don't know if I like that in highcolor toolbars...)

Still disabled because it only works with displays > 8bpp. The KStyle loader
will choose between this and the lowcolor one depending on display depth.

svn path=/trunk/kdelibs/; revision=46953