New style stuff. The old B2 will be removed in a sec.
I *think* I was able to hack this to be binary compatible except with
kdelibs/kstyles. Not sure tho ;-)

New features include highlighting the button when the mouse moves over it and
brush optimizations. Try moving the mouse around KCalc :)

I was able to reproduce the small scrollbar error by hacking a Qt example, but
like B3 the paintings still messed up so it's not in yet. I'll fix this
shortly (not a critical deal as it doesn't show up in apps, only some HTML).

Includes stubs for Kicker, but they don't do anything yet.

svn path=/trunk/kdelibs/; revision=47068