I was just about to say how hard-coding the height of the comboboxes in the toolbar to 30 is a Bad Idea.. when I remembered that it was *me* that hardcoded them to 20 before (which is a much better default, btw).  Anyway, it's better to have them a little dynamic.. so now the height is calculated by the font size plus some padding.  Try a 48 pixel font if you don't believe me.. hehe :-)
While I was munging around, I got rid of a few depreceated functions.
It appears that the setFullWidth function was used a bit.

I also see that enableFloating is marked as deprecated with a note to
use enableMoving.  HOWEVER, enableMoving doesn't quite take over the
entire role of enableFloating, so we can't get rid of enableFloating
that easily (yet).  I'll look at that later..

svn path=/trunk/kdelibs/; revision=47122