Pass the name of the desktop file to KRun::run (from KDEDesktopMimeType::runApplication) This should give more info to KRun::run (the XClassHintResName and NoAppStartNotify)) Hmm, Rikkus, what's the business with _dot_desktop, set and unused ?? Why isn't _desktop_file used to open the KDesktopFile ?
The problem with using kapp->startService() is that it doesn't support
command line options nor multiple arguments. We can use it only when
there is no args or just one URL. Still, this is the most common case.
Don't know how to integrate it with app-notification since there is no pid
in this case... Rikkus, over to you ;-)

svn path=/trunk/kdelibs/; revision=47477