Ok, ok, stop asking for it :)
Implemented KDirSize as a Job (to give it a nice API) that has one subjob
(a listRecursive). Synchronous call provided too.

Added it in KPropsDlg, _asynchronously_. This means, it pops up with "..."
and the "..." get replaced with the total dir size once we know it.
Even better than Microsoft's implementation (where you never know when it
has finished computing the size) ;-)

Note: works a lot faster after disabling debug output :)
Note 2: if you do it on '/', you'll never get the value, but closing kills
the job, as expected.

Final note: this is in kfile because we need it here. If you want me to
move it to KIO, no problem.

svn path=/trunk/kdelibs/; revision=49728