Well, I guess I don't know what was going on.  The Base64 encoding code was *wrong* all along... yet it hadn't been fixed and nobody was complaining.  This leads me to believe that either:
a) most web servers are *very* lax at decoding base64
or
b) most people have usernames and passwords such that their sizes plus
   one is divisible by 3
or
c) most sites that use authentication don't use basic

In any event, base64 requires the final string length to be divisible by
4 -- each 3 bytes (24 bits) of the original string becomes 4 bytes in
the base64 encoded version.  If the encoding results in a string that
is not divisible by 4, then the '=' character must be used as padding.

So it does now.

This allows me to authenticate with Zope (which is pretty strict) and
shouldn't break anything else UNLESS some web servers just plain decode
base64 wrong (I don't think so).

svn path=/trunk/kdelibs/; revision=50029