Linking change:
-libkmapnotify_la_LIBADD = $(LIB_DCOP) $(LIB_QT) ${LIBDL}
+libkmapnotify_la_LIBADD = ../dcop/libDCOP.la

It appears that on some system (mine, at least), if you explicitly
include -lqt or $(LIB_QT), then the linker will pick up libqt.so.1 as
a dependency (if it exists).  If you trust libtool to handle the
dependencies on it's own, though, everything links just fine.

In any event, all of the other libs in kdelibs use LIBADDs pointing
directly to the deps .la file so that also makes it more consistent.

svn path=/trunk/kdelibs/; revision=50387