The http slave should now do proper pattern matching on the user set UserAgents.  Barring bugs in my code, it should work as expected now.
However, I'm not a huge fan of the way I implemented it.  I use
QRegExp which is *very* handy.. but I'm guessing that it's also pretty
CPU intensive.  And since every single HTTP request (even for icons)
uses this function, I'm guessing that this could slow down things
considerably on slower machines.

I tried to alleviate this a little by storing a static hostname in the
function and if the current m_state.hostname is the same, it just
returns the last user_agent string.  This only works if the HTTP
requests come in order and to the same slave.

I'm not entirely sure how to do that more effeciently, though.

svn path=/trunk/kdelibs/; revision=50454