Use the KCursor::handCursor instead of the Qt hand cursor with KTextBrowser.  This is for consistency.
There is a little bit of flicker when it switches to the hand, though.
That's because it first switches to the Qt version in the
QTextBrowser code and I use that in the KTextBrowser version to know
when to switch to the hand.  I couldn't figure out any other way to do
this without completely reimplementing all of the enter events -- what
a pain.

svn path=/trunk/kdelibs/; revision=50746