Only two files, but many many changes :)
Moved back totalSize, processedSize and speed to Job, fixed the timer
that removes speed info after 5s, fixed progress info in many cases
(like FileCopyJob), moved the percent calculation code to Job::emitPercent,
etc.
This appears to be bin compat since those changes only affect the job classes,
not job.h - and apps only use jobclasses.h for the signals.

FileCopyJob still misses a gettingFile signal (which should be moved out
of the slaves), but I wonder if we shouldn't go for sending an infomessage
to the uiserver instead ?

svn path=/trunk/kdelibs/; revision=50817