Added an explicit 'updateRects' to updateGUI().
This is a bit confusing to me, though.  It is definitely needed or
else when the gui is rebuilt (from KEditToolbar, for instance), the
toolbars are undefined and the app will segfault.  HOWEVER, the
toolbar editor used to work just great without it!  Why?  What
changed?

Hmm... anyway, it seems to work (again) now.

svn path=/trunk/kdelibs/; revision=51252