Converted KToolBarItemList from an inherited QList to a typedef QValueList.  This was a bigger change then I thought it would be when I started.
Anyway, this all started when I was investigating why the bookmark
toolbar got all screwed up whenever you added new bookmarks (anywhere,
not just to the toolbar).  I tracked it down to the fact that the
KToolBar::clear() item in fact *wasn't* clearing everything.

So I got to thinking that a ValueList would be better suited here
anyway and since there is a QValueList::clear() method, it was
perfect!

However, I didn't realize just how much KToolBarItemList was used in
ktoolbar!  It feels like I had to change every other line of code...
but once you are 100 lines of changes in, you don't want to revert
everything and start over, so I finished

Technically, this *is* binary incompatible as the type of
KToolBarItemList has changed.  However, in practice, that list isn't
used anywhere but within the toolbar so it shouldn't be a big deal.

End result: the bookmark toolbar works correctly now :-)

svn path=/trunk/kdelibs/; revision=51668