implemented "level" which currently doesn't do everything
the point is to (almost) eliminate the use of KMessageBox (directly)
and instead, use KNotifyClient to send out error messages.  Certain error
messages, of course, are more important than others.

For example, "Notification" may be "There is a stain on the monitor"
while "Catastrophe" can be "There was a nuclear explosion nearby.  You have
1*10^-34 seconds to go into a box of half meter steel walls."

I'll have a commit for KNotifyClient and KAudioPlayer.  This shouldn't have
any source-level issues anywhere.

svn path=/trunk/kdelibs/; revision=51940