This should allow *most* Zope management pages to work with kio_http (not all.. yet).
The problem was due to the way that kio_http handled redirections.  I
don't know if this was really our problem or if Zope does it wrong....
but I'm thinking it's ours.

This all centers on the m_iSize variable.  By default, it is set to 0
and is later set according to the 'Content-length' field.  Later on,
in the ::readBody() function, there is the code (essentially):

if (m_iSize)
  readLimited();
else
  readUnlimited();

That is, m_iSize is sorta being used a boolean here to mean "do I have
Content-length?" as well as the actual size.

However, in the case of redirections, the content-length CAN be 0,
meaning that there is *no* content -- just the header.  Unfortunately,
with the code as is, this means that readUnlimited() will be called.
This creates some undefined behavior..

SO.. I changed the default setting of m_iSize to -1.  Any resetting
sets it back to -1.  The only way it will be 0 is if Content-length is
0.

I tested this with a number of sites so I don't think it should cause
any harmful effects.  Let me know if I'm wrong!

svn path=/trunk/kdelibs/; revision=52192