Did you sleep well, Stephan? Are you better?
OK: first SORRY to have done this a day before a freeze. I did not
remember about this freeze.

Second, we NEED to have a copy of the DocBook tools hosted at KDE. I
cannot spend so much time on debugging this, neither can Mike, so those
changes are both NEEDED and IMPORTANT.

Of course, this should work with the old setups and without the KDE tools
as well. I will try to fix that. I did not try to ruin your day ;-).

This pecular change fixes the call to install-catalog.

svn path=/trunk/kdelibs/; revision=52261