This is a rather big change to the way KHTML handles forms.
Background
----------
Until now, KHTML did not handle multipart forms at all.  That is, it
couldn't handle those forms with ENCTYPE="multipart/form-data".  In
pratical terms, this means those forms that have a "Browse..." button
for uploading files.  They are very common in web-admin tools and
on-line image manipulation sites and "build your web-page online"
tools.

Implementation
--------------
Implementing this was quite a bit more work then I thought it would
be.  The core problem was that it is possible to "upload" *binary*
files (like gifs or pngs or whatever) inside of the form.  However, up
until now, the form data was stored in a QString.  Specifically, each
form element returned it's data in the 'QString encoded()' function.

Unfortunately, QStrings do not like data that has NULL characters in
it.  NULL chars are *very* common in images, so this was not good.

In the end, I had to change the 'encoded()' function to return
QCString instead of QString.  This is because QCString is just a
glorified QByteArray... and if you used QCString::size() instead of
QCString::length(), it would return the correct size even if it's
storing a binary image.

I hit another snag when I realized that kio_http always assumed
'Content-type: application/x-www-form-urlencoded'.  I modified
kio_http to handle other headers... but how to pass that along?

I ended up modifying the KHTMLPart::submitForm function to accept
those params.

Problems
--------
This is quite a bit of a hack, but it does appear to work!  However,
there are some *potential* problems.

First, the use of QCString instead of QString in the encoded()
functions mean that it's probably no longer possible to use UNICODE
strings in the forms.  I committed it this way since I'm not sure
that you can do it *anyway*.  If I'm wrong, then we somehow have to
figure out how to store binary data in a QString

Second, I had to change a decent number of function signatures.  I am
99.999% sure that all functions changed are internal functions... but
if any of them *are* used external to khtml, then this is a binary
incompatible change.  Like I said, I'm almost 100% sure that they
*are* all internal since I could run konqueror with no recompiles at
all.

Let me know if there are any other problems.

svn path=/trunk/kdelibs/; revision=52773