Corrected some size problems.
The IconList was in certain cases too narrow.
The reason for this behavior is that I use a
bold font for the text but I did not compute the width
of the list items using the bold font.

The size calculation of the IconList got a bit broken
when the new features were added a few days ago. I have
added a couple a new functions:
KJanusWidget::IconListBox::invalidateWidth() and
KJanusWidget::IconListBox::updateWidth to make the size
calculation work again. It is more efficient and simpler
to understand.

svn path=/trunk/kdelibs/; revision=52901