Implemented Waldo's idea. KApplication holds a reference counter: * This counter indicates when to exit the application. * It starts at 1, is decremented by the "last window close" event, but * is incremented by operations that should outlive the last window closed * (e.g. a file copy for a file manager, or 'compacting folders on exit' for a mail client).
This fixes "konqueror aborts a save as operation when closing the last window"
(#4561, from Waldo as well :-) and could be useful to e.g. kmail.

Note: QApplication's lastWindowClosed signal breaks this. BUT KTMainWindow
was already doing the app->quit on last window, so it was already useless.
Do NOT use QApplication's lastWindowClosed.

The change is binary compatible, long live d pointers.

svn path=/trunk/kdelibs/; revision=54249