Thrashed binary compatibility;
I'm adding a paintEvent() method so that when an icon is dragged over, and a line
is drawn, and the icon is dragged over the line, then the line isn't erased.

I havn't actually implemented it yet because I'm lazy :)  But I'll get it tomorrow,
which means today, but I still like my timezone.  /me rambles.

svn path=/trunk/kdelibs/; revision=55457