Rewrote KURLLabel. It's 99.9% source-compatible, the remaining two methods aren't used anywhere as far as I know.
Also it should be binary compatible in practice: QWidgets are/should be
always allocated on the heap, and various setText/Pixmap/Movie functions
were virtual anyway.

Furthermore, I don't think anyone derived from it, class layout is not a
problem in this regard.

Special "handling" code for bugs in the old KURLLabel (for example in
kbugreport.cpp) should be removed (I'll do that for KBugReport, and any
other uses of KURLLabel that I come across :).

svn path=/trunk/kdelibs/; revision=56672