Added new debugging code for internal latencies (IOMANAGER_DEBUG_LATENCY):
Enable this if you want to debug how long certain plugins / operations
take to perform. You'll get the times between two select() calls that are
done by the IOManager, which is equivalent to the time the input/output
remains unserved. For apps like artsd, it gives the minimum audio latency
users will need to specify to avoid dropouts.

svn path=/trunk/kdelibs/; revision=57953