Ok, this is a change in the signature of a global function, the function was declared as UserIconSet(QString, int, KInstance) but was implemented (some days before the freeze) as UserIconSet(QString, KInstance)
This wasn't used anywhere, that's why we didn't had any problem with it.

mhk: Note that when talking to you yesterday, I thought that I had to
change also the implementation, but I noticed today that the implementation
was ok since before the freeze.

Conclusion: this is a binary compatible change.

svn path=/trunk/kdelibs/; revision=58224