Fixed editing a file over FTP (bug #xyzt and reminded by Andreas Zehender) :
KIO::NetAcces::upload should in fact set overwrite to true, since this is
about saving the changes to an existing file. Well, distinguishing between
save and save as is difficult here. But an app can use KIO::NetAccess::exists
to popup a confirmation. Better than what we had up to now: an error 12 in the
background and no way to set overwrite.

netaccess.h: updated docu about this.

svn path=/trunk/kdelibs/; revision=58272