Add link overlay for links and padlock overlay for non-readable files (TODO: fix the latter, at the moment it only works with u-r/g-r/o-r files).
Not a new feature, simply using KIconLoader's overlays :)
(and asked by Torsten and in #6555, and those who report that deleting
a link to a dir deletes the pointed dir - it does not but it seems so !)

Now, I fear I'm going to hear "this is too windowsish and I want to turn
it off...". Well. Even I got confused by links that didn't look like links :-]

svn path=/trunk/kdelibs/; revision=58609