Deselect the current item when it's being executed (as opposed to being selected).
Here's my rationale for this:

When I was visiting my family recently, I had several of my (Windows95
using) brothers and sisters try out KDE2 and *all* of them did the
following when trying to start something from the desktop:

a) Double-click the icon
b) Notice that the icon was selected so obviously that means that the
   double-click failed
c) Double-click again

This results in four instances of whatever it was that they were
starting.

Now ignoring for the minute that the app-starting notification and
zooming rectangle were both *completely* lost on them, it's clear that
they are used to the windows paradigm where the icon is deselected when
it's executed.  Since they are probably pretty typical, I thought this
is something we should honor as well.

svn path=/trunk/kdelibs/; revision=59137