Fixed two bugs... one very badly.
Bug 1:
If  no OPTION is 'selected' in a SELECT, then the encoding failed and
nothing was returned.  This will return the first one.

Now this behavior is actually undefined.. the spec says that the code
should always have one pre-selected option.  However, RFC1886 says to
go with the first ones and several existing web pages obviously assume
that.

Bug 2:
This is fixed in the wrong place.  Apparently, the label of the option
sometimes has a space at the end of it.  The big problem with this is
that the encoding will treat this like:
  myOption=Value+

I just did a simplifyWhiteSpace() on it.  However, this is the wrong
place to fix it... the getAttribute(ATTR_LABEL) in ::text() shouldn't
have the space in the first place!

svn path=/trunk/kdelibs/; revision=59342