Applied patch by Kevin Puetz (who wrote it for Powermac compatibility).
It verifies after opening /dev/dsp whether the fragment count/size could
be supplied by the sound card, and if not modifies the aRts settings to
match the real (supplied) soundcard settings.

This should make artsd start for those who have now seen it stopping with
an assert in audiosubsys.cc.

svn path=/trunk/kdelibs/; revision=62428